dom/base/test/test_document.all_iteration.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test for document.all iteration behavior</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_array_equals([...document.all], document.getElementsByTagName("*"));
}, "document.all should be iterable");
</script>
